requirements.txt
aniso8601==8.1.0
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect2==0.4.1
defusedxml==0.6.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.1.0
pycparser==2.20
python-decouple==3.4
Pyphen==0.10.0
pytz==2020.5
requests==2.25.1
six==1.15.0
tinycss2==1.1.0
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
Werkzeug==1.0.1
